backend/requirements.txt
aioredis
amqp
asgiref
asttokens
async-timeout
attrs
autobahn
Automat
billiard
cffi
click
click-didyoumean
click-plugins
click-repl
colorama
coverage
constantly
cryptography
channels
channels_redis
daphne
Django<4
django-channels-graphql-ws
django-graphql-jwt==0.3.0
django-graphql-auth
django-cors-headers
django-grappelli
django-redis
django-timezone-field
django-cleanup
gunicorn
graphene-django
et-xmlfile
executing
hiredis
hyperlink
icecream
idna
incremental
kombu
memoize
msgpack
numpy
openpyxl
pandas
prompt-toolkit
psycopg2
pyasn1
pyasn1-modules
pycparser
Pygments
pyOpenSSL
pyparsing
python-crontab
python-dateutil
pytz
redis
rq
scipy
service-identity
six
sqlparse
Twisted
txaio
times
vine
wcwidth
zope.interface
whitenoise